requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.8
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
future==0.18.2
importlib-metadata==1.7.0
inflection==0.5.0
kombu==4.6.11
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
simplejson==3.17.2
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
